refactor(thumbnail): サムネイルコアロジックをcore_logicクレートに移動
## 変更内容
- config, error, generator, batch, folder, utils(hash_path)をcore_logic/src/thumbnail/に移動
- thumbnail.rsをTauriコマンドの薄いラッパーに変更
- get_cache_dirのみTauriコマンド層に残留

## 技術的変更
- core_logic/Cargo.toml: image, rayon, blake3, thiserror, num_cpus, serde_json追加
- src-tauri/Cargo.toml: サムネイル専用依存を削除（core_logic経由に）
- core_logic/src/thumbnail/: 新規モジュール（6ファイル + mod.rs）
- src-tauri/src/commands/thumbnail/: ディレクトリ削除
- src-tauri/src/commands/thumbnail.rs: core_logic::thumbnail からの import に変更

## 検証結果
- cargo test -p core_logic: 19 passed
- cargo test: 全テストPASS
