+++ src-tauri/Cargo.toml
@@ -32,13 +32,6 @@ tauri-plugin-fs = "2"
 tauri-plugin-store = "2"
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
 
-# Thumbnail optimization dependencies (001-rust-thumbnail-optimization)
-image = { version = "0.25", default-features = false, features = ["png", "jpeg", "webp"] }
-rayon = "1.11"
-blake3 = "1.8"
-thiserror = "2.0"
-num_cpus = "1.16"
-
 [profile.dev]
 debug = true
 opt-level = 0
