+++ src-tauri/core_logic/Cargo.toml
@@ -5,3 +5,9 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1"
+image = { version = "0.25", default-features = false, features = ["png", "jpeg", "webp"] }
+rayon = "1.11"
+blake3 = "1.8"
+thiserror = "2.0"
+num_cpus = "1.16"
