+++ src-tauri/core_logic/src/thumbnail/config.rs
