+++ src-tauri/core_logic/src/thumbnail/error.rs
