+++ src-tauri/core_logic/src/thumbnail/generator.rs
@@ -1,8 +1,8 @@
 // サムネイル画像生成のコアロジック
 
-use super::config::ThumbnailConfig;
-use super::error::{Result, ThumbnailError};
-use super::utils::hash_path;
+use crate::thumbnail::config::ThumbnailConfig;
+use crate::thumbnail::error::{Result, ThumbnailError};
+use crate::thumbnail::utils::hash_path;
 use image::imageops::FilterType;
 use image::{GenericImageView, ImageFormat};
 use std::path::{Path, PathBuf};
