+++ src-tauri/core_logic/src/thumbnail/mod.rs
@@ -1,0 +1,13 @@
+pub mod batch;
+pub mod config;
+pub mod error;
+pub mod folder;
+pub mod generator;
+pub mod utils;
+
+pub use batch::{BatchResult, BatchTask, BatchThumbnailGenerator, TaskPriority};
+pub use config::ThumbnailConfig;
+pub use error::{Result, ThumbnailError};
+pub use folder::FolderThumbnailResult;
+pub use generator::ThumbnailGenerator;
+pub use utils::hash_path;
