+++ src-tauri/src/commands/thumbnail.rs
@@ -1,29 +1,30 @@
 // サムネイル生成と管理のためのTauriコマンド
+//
+// コアロジックは core_logic::thumbnail に移動済み。
+// このファイルは Tauri コマンドの薄いラッパーとして、以下のみを担当:
+// 1. AppHandle → PathBuf の変換（get_cache_dir）
+// 2. core_logic::thumbnail::* の呼び出し
+// 3. 結果の Tauri IPC 向けシリアライズ
 
-mod batch;
-mod config;
-mod error;
-mod folder;
-mod generator;
-mod utils;
-
-pub use batch::{BatchResult, BatchTask, BatchThumbnailGenerator, TaskPriority};
-pub use config::ThumbnailConfig;
-pub use error::{Result, ThumbnailError};
-pub use folder::FolderThumbnailResult;
-pub use generator::ThumbnailGenerator;
-pub use utils::{get_cache_dir, hash_path};
-
+use core_logic::thumbnail::folder;
+use core_logic::thumbnail::{
+    BatchTask, BatchThumbnailGenerator, FolderThumbnailResult, ThumbnailGenerator,
+};
 use tauri::command;
 
+/// サムネイルキャッシュディレクトリのパスを取得（Tauri依存）
+fn get_cache_dir(app_handle: &tauri::AppHandle) -> std::io::Result<std::path::PathBuf> {
+    use tauri::Manager;
+    let cache_dir = app_handle
+        .path()
+        .app_cache_dir()
+        .map_err(|e| std::io::Error::new(std::io::ErrorKind::NotFound, e))?;
+    let thumbnail_dir = cache_dir.join("thumbnails");
+    std::fs::create_dir_all(&thumbnail_dir)?;
+    Ok(thumbnail_dir)
+}
+
 /// 画像のサムネイルを取得または生成する
-///
-/// # Arguments
-/// * `image_path` - ソース画像のフルパス
-/// * `app_handle` - Tauriアプリケーションハンドル（自動注入）
-///
-/// # Returns
-/// サムネイルのキャッシュパス（成功時）
 #[allow(dead_code)]
 #[command]
 pub async fn get_or_create_thumbnail(
@@ -33,28 +34,17 @@ pub async fn get_or_create_thumbnail(
     let cache_dir = get_cache_dir(&app_handle).map_err(|e| e.to_string())?;
     let generator =
         ThumbnailGenerator::with_default_config(cache_dir).map_err(|e| e.to_string())?;
-
     let cache_path = generator
         .get_or_create_thumbnail(&image_path)
         .map_err(|e| e.to_string())?;
-
     let path_str = cache_path
         .to_str()
         .map(|s| s.to_string())
         .ok_or_else(|| "Failed to convert path to string".to_string())?;
-
     Ok(path_str)
 }
 
 /// 複数の画像のサムネイルをバッチ生成する
-///
-/// # Arguments
-/// * `image_paths` - ソース画像のパスの配列
-/// * `visible_count` - 可視領域の画像数（優先度High）
-/// * `app_handle` - Tauriアプリケーションハンドル（自動注入）
-///
-/// # Returns
-/// 各画像パスに対応するサムネイルキャッシュパスまたはエラーメッセージのマップ
 #[allow(dead_code)]
 #[command]
 pub async fn batch_create_thumbnails(
@@ -62,34 +52,27 @@ pub async fn batch_create_thumbnails(
     visible_count: Option<usize>,
     app_handle: tauri::AppHandle,
 ) -> std::result::Result<std::collections::HashMap<String, serde_json::Value>, String> {
-    // tokio::spawn_blockingで非同期に実行（UIをブロックしない）
     let cache_dir = get_cache_dir(&app_handle).map_err(|e| e.to_string())?;
     tokio::task::spawn_blocking(move || {
-        // BatchThumbnailGeneratorの初期化
+        use core_logic::thumbnail::TaskPriority;
         let batch_generator =
             BatchThumbnailGenerator::with_default_config(cache_dir).map_err(|e| e.to_string())?;
-
-        // タスクリストの作成（優先度付き）
         let visible_count = visible_count.unwrap_or(10);
         let tasks: Vec<BatchTask> = image_paths
             .into_iter()
             .enumerate()
             .map(|(index, image_path)| {
                 let priority = if index < visible_count {
-                    TaskPriority::High // 可視領域
+                    TaskPriority::High
                 } else if index < visible_count * 3 {
-                    TaskPriority::Normal // 近傍
+                    TaskPriority::Normal
                 } else {
-                    TaskPriority::Low // バックグラウンド
+                    TaskPriority::Low
                 };
                 BatchTask::new(image_path, priority)
             })
             .collect();
-
-        // バッチ生成の実行
         let results = batch_generator.batch_create_thumbnails(tasks);
-
-        // 結果をマップに変換
         let result_map: std::collections::HashMap<String, serde_json::Value> = results
             .into_iter()
             .map(|result| {
@@ -112,7 +95,6 @@ pub async fn batch_create_thumbnails(
                 (result.image_path, value)
             })
             .collect();
-
         Ok(result_map)
     })
     .await
@@ -122,58 +104,37 @@ pub async fn batch_create_thumbnails(
 /// サムネイルキャッシュをクリアする（デバッグ用）
 #[command]
 pub async fn clear_thumbnail_cache() -> std::result::Result<(), String> {
-    // TODO: Phase 6でキャッシュ管理を実装
     Err("Not implemented yet".to_string())
 }
 
-/// フォルダのサムネイルを取得する（フォルダパスのみで完結）
-///
-/// フォルダ内の最初の画像を自動選択し、サムネイルを生成して返す。
-/// 画像がないフォルダの場合は None (null) を返す。
-///
-/// # Arguments
-/// * `folder_path` - 対象フォルダのフルパス
-/// * `app_handle` - Tauriアプリケーションハンドル（自動注入）
-///
-/// # Returns
-/// FolderThumbnailResult（画像パス、サムネイルパス、ファイル名）、画像なしの場合 null
+/// フォルダのサムネイルを取得する
 #[command]
 pub async fn get_folder_thumbnail(
     folder_path: String,
     app_handle: tauri::AppHandle,
-) -> std::result::Result<Option<folder::FolderThumbnailResult>, String> {
+) -> std::result::Result<Option<FolderThumbnailResult>, String> {
     let cache_dir = get_cache_dir(&app_handle).map_err(|e| e.to_string())?;
     let result = tokio::task::spawn_blocking(move || {
-        // 1. フォルダ内の最初の画像を取得
         let first_image = folder::get_first_image_in_folder(&folder_path)?;
-
         let image_path = match first_image {
             Some(path) => path,
-            None => return Ok::<Option<folder::FolderThumbnailResult>, String>(None),
+            None => return Ok::<Option<FolderThumbnailResult>, String>(None),
         };
-
-        // 2. サムネイルを生成
         let generator =
             ThumbnailGenerator::with_default_config(cache_dir).map_err(|e| e.to_string())?;
         let cache_path = generator
             .get_or_create_thumbnail(&image_path)
             .map_err(|e| e.to_string())?;
-
-        // 3. basenameを取得
-        // get_first_image_in_folder が画像ファイルパスのみ返すため
-        // file_name() が None になるケースは実質到達しない（防御的フォールバック）
         let image_name = std::path::Path::new(&image_path)
             .file_name()
             .and_then(|n| n.to_str())
             .unwrap_or("unknown")
             .to_string();
-
         let thumbnail_path = cache_path
             .to_str()
             .map(|s| s.to_string())
             .ok_or_else(|| "Failed to convert thumbnail path to string".to_string())?;
-
-        Ok(Some(folder::FolderThumbnailResult {
+        Ok(Some(FolderThumbnailResult {
             image_path,
             thumbnail_path,
             image_name,
@@ -181,19 +142,10 @@ pub async fn get_folder_thumbnail(
     })
     .await
     .map_err(|e| format!("Task join error: {}", e))??;
-
     Ok(result)
 }
 
-/// 複数フォルダのサムネイルをバックグラウンドでプリフェッチする
-///
-/// フォルダパスの配列を受け取り、各フォルダの最初の画像のサムネイルを
-/// バッチ生成する。優先度はインデックス順で自動決定される。
-/// Fire-and-forget パターンで使用される想定。
-///
-/// # Arguments
-/// * `folder_paths` - 対象フォルダのパス配列
-/// * `app_handle` - Tauriアプリケーションハンドル（自動注入）
+/// 複数フォルダのサムネイルをプリフェッチする
 #[command]
 pub async fn prefetch_folder_thumbnails(
     folder_paths: Vec<String>,
@@ -201,7 +153,6 @@ pub async fn prefetch_folder_thumbnails(
 ) -> std::result::Result<(), String> {
     let cache_dir = get_cache_dir(&app_handle).map_err(|e| e.to_string())?;
     tokio::task::spawn_blocking(move || {
-        // 1. 各フォルダから最初の画像パスを取得
         let image_entries: Vec<(usize, String)> = folder_paths
             .iter()
             .enumerate()
@@ -212,12 +163,9 @@ pub async fn prefetch_folder_thumbnails(
                     .map(|image_path| (index, image_path))
             })
             .collect();
-
         if image_entries.is_empty() {
             return Ok(());
         }
-
-        // 2. 優先度付きバッチタスクを作成
         let tasks: Vec<BatchTask> = image_entries
             .into_iter()
             .map(|(index, image_path)| {
@@ -225,12 +173,9 @@ pub async fn prefetch_folder_thumbnails(
                 BatchTask::new(image_path, priority)
             })
             .collect();
-
-        // 3. バッチ生成を実行
         let batch_generator =
             BatchThumbnailGenerator::with_default_config(cache_dir).map_err(|e| e.to_string())?;
         let _results = batch_generator.batch_create_thumbnails(tasks);
-
         Ok(())
     })
     .await
